Fixed an AnaccessViolation exception due to feed objects.
ReportData was freed before updating causing the subreports to be freed
to. ReportData was freed bcause the UpdateFunction was using a create
constructor to put the new data and therefore the old wasn't needed.
But as the cleaning of data was improved and included freeing of
subreports to. The objects in the listbox was still poining to the old
subreports that no longer exists and caused an Access Violation whenthey
where refered.

Now the update function are replaced by a procedure that takes the
ReportData as a parameter and instead of calling the TCMMReportData.Create
it calls the object using a new Update procedure.